+++ src/api/user.py
@@ -110,72 +110,3 @@ def kakao_unlink(request: KakaoUnlinkRequest, db: Session = Depends(get_db)):
             # Default to 500 for unexpected errors
             raise HTTPException(status_code=500, detail=f"Internal server error: {str(e)}")
 
-
-"""
-아래 코드는 카카오 로그인 테스트 용으로 사용하는 코드입니다.
-실제 로그인 시 사용하지 않습니다. (프론트 쪽에서 처리)
-"""
-# @router.post("/kakao/get-real-code")
-# async def get_real_kakao_authorization_code(request: dict, db: Session = Depends(get_db)):
-#     """
-#     Get REAL Kakao authorization code using email and password.
-#     This calls the actual Kakao API.
-#     """
-#     try:
-#         email = request.get("email")
-#         password = request.get("password")
-        
-#         if not email or not password:
-#             raise HTTPException(status_code=400, detail="Email and password are required")
-        
-#         logger.info(f"Attempting to get real Kakao authorization code for: {email}")
-        
-#         # Call Kakao's login API to get authorization code
-#         # This is the real Kakao authentication flow
-#         import requests
-        
-#         # Kakao login endpoint
-#         login_url = "https://accounts.kakao.com/login"
-        
-#         # Create session to maintain cookies
-#         session = requests.Session()
-        
-#         # First, get the login page to get any required tokens
-#         login_page_response = session.get(login_url)
-        
-#         # Now attempt to login with credentials
-#         login_data = {
-#             "email": email,
-#             "password": password,
-#             "continue": "https://kauth.kakao.com/oauth/authorize"
-#         }
-        
-#         login_response = session.post(login_url, data=login_data)
-        
-#         # Check if login was successful
-#         if login_response.status_code == 200 and "authorization_code" in login_response.text:
-#             # Extract authorization code from response
-#             import re
-#             code_match = re.search(r'authorization_code=([a-zA-Z0-9]+)', login_response.text)
-#             if code_match:
-#                 authorization_code = code_match.group(1)
-                
-#                 return {
-#                     "authorization_code": authorization_code,
-#                     "email": email,
-#                     "name": email.split("@")[0],
-#                     "message": "Real authorization code obtained from Kakao"
-#                 }
-        
-#         # If we get here, login failed
-#         raise HTTPException(status_code=401, detail="Invalid Kakao credentials")
-        
-#     except HTTPException:
-#         raise
-#     except Exception as e:
-#         logger.error(f"Error getting real Kakao authorization code: {str(e)}")
-#         raise HTTPException(status_code=500, detail=f"Failed to get authorization code: {str(e)}")
-
-
-
-
